+++ internal/cmd/logs/logs.go
@@ -61,6 +61,6 @@ func init() {
 	Cmd.Flags().BoolVarP(&allLogs, "all", "a", false, "include uninteresting logs")
 	Cmd.Flags().BoolVarP(&chart, "chart", "c", false, "Output a gantt chart of the most recent action as an html file")
 	Cmd.Flags().BoolVar(&config.Debug, "debug", false, "Output debugging information")
-	Cmd.Flags().IntVarP(&logsLength, "length", "l", 0, "limit the number of logs starting from the latest")
-	Cmd.Flags().IntVarP(&logsDays, "days", "d", 0, "limit the number of logs starting from the latest")
+	Cmd.Flags().IntVarP(&logsLength, "length", "l", 0, "Number of logs to display")
+	Cmd.Flags().IntVarP(&logsDays, "days", "d", 0, "Age of the logs to display in days")
 }
